requirements.txt
pytest>=8.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
httpx>=0.26.0
pydantic>=2.6.0,<3.0.0
scikit-learn>=1.3.0
pyyaml>=6.0.0
Pillow>=10.0.0
openai>=1.40.0,<2.0.0
psycopg2-binary>=2.9.9
